package refactoring -> models.auth
