+++ app/models/Member.java
@@ -1,5 +1,6 @@
 package models;
 
+import models.auth.AuthAccount;
 import com.google.common.base.Function;
 import com.google.common.base.Predicate;
 import com.google.common.collect.Collections2;
