+++ app/models/auth/TwitterOAuthAccount.java
@@ -1,6 +1,9 @@
-package models;
+package models.auth;
 
 import javax.persistence.Entity;
+import models.Member;
+import models.ProviderType;
+import models.TwitterAccount;
 import org.apache.commons.lang.StringUtils;
 import play.data.validation.Required;
